splict random_test.cpp into run_method it include how method are called
